requirements_min.txt
pandas
streamlit
langchain-openai
langchain-community
langchain-text-splitters
langgraph
pypdf